Main push
In this push I started fresh from the beginning. I stripped back the project and rebuilt it.

In rebuilding i was able to add websockets based off of content provided on Socket.io as well as starting to implement mongoose DB.

Overall at this point there is still more required in styling the site and adding more content in